services/kafka-producer/requirements.txt
kafka-python==2.0.2
aiohttp==3.9.1
asyncio-mqtt==0.13.0